Addressing comments by copilot
